Initial import of project.
Status of the various parts

Parts of the gui are outlined. 
The list of idd objects is created and displayed in a tree.
The class generation from the idd file is almost complete. There are some specific classes that need some fixups.
The model for active objects works and treeview displays the objects.
Reading from an idf works partially. There are some classes that don't read properly, and it is barely tested.

To do:
Finish the fixups on class generation. test class generation with other than v 3.1 idd.
Write the idf writing code, taking the items from the model and writing. Test on various sample idfs for accuracy.
Write the gui for editing classes.