Filling out idfabstractmodel.
Implemented data() method.

Added a role to idfglobals.py for data() to return class instance

Implemented columnCount(). This is preliminary, and I'll see how it works out in usage. If there is no parent modelindex passed, the columnCount() returns 2. The two columns are iddclassname and classname. If there is a parent passed, then columnCount() returns the fieldcount for the object. The structure the model presents is a tree the classname and iddclassname as parents, and all the other fields as children.

Added some methods to objectclass to retrieve necessary information.