Reworked fieldclasses and some widgets to allow editing. The simple ones work, ie text edit, spin boxes and combo boxes.
Unfinished: object combo boxes, ie in surface:detailed the zone. Need lists of types of classes to fill the combo boxes.

Schedules don't work. Nor do the vertice editing.

Saving to the model isn't implemented yet.

The design of the extensible classes is in question. Currently the last field of the extensible list has all the data.

The extensible classes have repeating field groups made up of any number of fields.

For example:
class WindowMaterial_GlazingGroup_Thermochromic(ObjectAbstract):
  def CreateFields(self) :
    self.InsertField(FieldText(self,"Name","",""))
    self.InsertField(FieldReal(self,"Optical Data Temperature 1",0,"","","","",""))
    self.InsertField(FieldObjectlist(self,"Window Material Glazing Name 1","","","GlazingMaterialName"))
...
   self.setExtensible(2)
...

The last two fields are repeating. 

When the data is loaded from the idf file, right now the data is included in the last field as restofdata.

We need easy mapping to model and table view  for editing.
A way to add records*extinsible count up to the limit defined in the idd.

Derek