Modified ObjectAbstract to handle extensible classes.
The idd class remains the same. When a class is extensible, a copy is made of the fields. When data is read from an idf file, the classes are instantiated, and if there is more data than fields, the set of extensible fields are created. This gets rid of the restofdata nonsense and simplifies displaying the class in a table.

The construction class with vertices needs to be changed.
An extensible class needs the ability to add or remove fields when editing. The model needs to be modified to handle those cases.
Writing out the class needs modification in extensible classes.

Added to the idfClassModel the ability to setData. Edits are saved now in the model.

Derek