Fixed bug in iddclassgen.py, where if the class was defined Wall:Exterior, /memo ..., the class name was not extracted properly. In those instances, the next defined class could be misdefined. An updated iddclass.py is included. Corrected the indentation of idfclassgen.py.
Added newclassdialog.py. A dialog that shows the list of all defined idd classes for the purpose of adding a new class to the idf definition. Currently it only shows a treeview. The treeview has as root a sorted list of all the groups, and each group presents a sorted list of classes.