Implemented newclassdialog. One can select one or more classes from the idd. The dialog returns a list of class names.
In the test app, the list of classnames are instantiated and inserted into the model.

Extracted and set as default the idd version to the Version class. Changed the class generator and class file.

Implemented a method insertRecord in idfdata. A class instance is passed, and it rebuilds the tree. The model must be reset to read the new data.

I'm thinking that we have a basic editor that can load, edit and save, or create new, edit and save. What do you think of doing a release of this basic functionality? After some testing of course. I intend to build a building simulation using this, and no doubt will find errors.

Derek