Fieldclasses.py - put try around loading object-list items from depends dict. Would throw an error if no reference classes were instantiated. May put up an error box notifying the user.
idfmodeltest.py - set the class  table to fill horizonally

idfmodeldelegate.py - added a call to the idfclass model data() with editrole.

idfabstractmodel.py - the call to data() with editrole set and if on the last row generates a call to insertRows() to insert extensible fields if object is extensible. Also changed rowCount() routine to return length of field list.