Cleaned up a bunch of debug prints
Fixed class memo and field notes so they show up as tool tips in the model views.

Compact:Schedule -  fixed editing. Fields are inserted for editing. If a Through: section is added, it inserts For: and Interpolate: sections. When the For: line is edited, Until: Data: pairs are inserted.

Editing works, bugs notwithstanding.