Added loadclassdialog.py.
Implements dialog for loading classes defined in an idf file.

The dialog can load multple idf files, and class objects can be selected. A textedit allows viewing the data of the object. If an object has dependancies, the depends can be copied as well. For example, if a Construction object is selected, which is made up of 3 Material objects, all four will be copied.

added a deleteRecord() method to idfdata.py.