idfdata.py: cleanup of file
graphicitem.py: beginnings of QGraphicsItem subclass of the polygon item. The intent is to read from the various building surface classes, getting either vertices or other size data to build a polygon that will be displayed in a graphicsscene. 5 views will be built, looking down, looking from north, south, east and west.