In iddclassgen.py fixed setting required object. Was being set if the field was required
Added code to set fields as required. Changed class generation to add a call to setRequiredFields, which takes a list of fields that are required by the simulation. Added a method to fieldclasses.py that sets the field as required.

Not used for now. Will be used by the surface element tree building routines to see if the object is defined or not.