Organizing the data structure for building elements. idfzonemodel.py is the data model that encapsulates a tree made up of:
Building - Zone1 -  Surface
	          - sub surface
	-  Surface
	-  ...
             - Zone2
             - ... 

The nodes are treeItems from idfdata.py. The data field is a surfaceitem which is passed the idfobject instance, and creates a shape instance and a graphics item instance.

When the tree is build, the shape class creates a vertice list that will be used by the graphicsitem to display the object.