Saving work. Refactored verticemath to make it into a stand along calculation class that can cache matrices for speed.
Moved most of the logic from verticemath into the graphicitems classes. Created a zoneItem class that can display either the outline of the zone, or it's children. The items build the vertice lists from the idf classes. 

Reworked the zonemodel class to support changes

added idfzoneview.py, which is a subclass of qgraphicsview. Will change into a widget that contains a view and all the editing and viewing and navigation widgets. Beginnings of a zorder class that will be used to calculate levels in a building, levels in a zone, which will be used to decide what items to show.

Not working as yet, just saving work.